Déplacement de la popup iframe dans un nouveau composant
